requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.56.1
django-extensions==3.2.3
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.0
idna==3.4
Markdown==3.4.4
msgpack==1.0.5
multidict==6.0.4
oauthlib==3.2.2
openai==0.28.0
Pillow==10.0.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
